Revert "2965 - migrate Property logic into the useProperty hook"
This reverts commit cb16eddc4dd9e4633b71b3db34923695781465a8.
